Remove -mnocygwin
The flag is useless as it is, since it should be -mno-cygwin. That being
said, fixing the flag prevent regsvr32 from registring the lib properly.

Since no one seems to have realized the flag was not working, it should be
safe to remove it.

This fixes https://github.com/wineasio/wineasio/issues/115

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>