Allow to use a different PKG_CONFIG_PATH when building 32bit dll/lib
On 32bit distros, the default /lib or /usr/lib are used to share libraries
needed for the system to work.

However, on 64bit distributions, things can take a few forms. Because we
may support multiple binary formats, a separation between native 64bit
libraries and 32bit ones is done. This could mean having:
/lib and /lib32
/lib, /lib64 and /lib32
/lib/supported-binary-format (on Debian-based distro, this would be
lib/x86_64-linux-gnu and lib/i386-linux-gnu, etc.)
and so on, all this with symlinks here and there to ensure a maximum
compatibility between distros.

For this reason, we should be able to tell pkgconfig where to look for
PC files to find either 64bit or 32bit libraries through the standard
PKG_CONFIG_PATH variable.

This patch add the PKG_CONFIG_PATH variable, bring in its value if it was
defined outside Makefile.mk and set it to /usr/lib$(M)/pkgconfig/ if it
wasn't set.

This also fixes a benign error where pkgconfig would return a library's
CFLAGS and installation folder for the 64bit format when building a 32bit
wineasio.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>